requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
backcall==0.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2
django-celery-beat==1.5.0
django-embed-video==1.2.0
django-extensions==2.2.3
django-filter==2.2.0
django-heroku==0.3.1
django-shell-plus==1.1.7
django-timezone-field==3.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
freeze==1.0.10
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
lxml==4.4.1
parsel==1.5.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pipenv==2018.11.26
prompt-toolkit==2.0.10
psycopg2==2.8.3
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-crontab==2.3.9
python-dateutil==2.8.0
python-decouple==3.1
python-scrapyd-api==2.1.2
pytz==2019.1
queuelib==1.5.0
requests==2.22.0
Scrapy==1.7.3
scrapyd==1.2.1
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.3
Twisted==19.7.0
urllib3==1.25.3
virtualenv==16.7.2
virtualenv-clone==0.5.3
w3lib==1.21.0
wcwidth==0.1.7
whitenoise==4.1.3
zope.interface==4.6.0
